refactor(peppylib): remove force_standalone, move builder tests to integration suite
